requirements.txt
langchain==0.3.13
langchain-google-genai==2.0.5
langchain-openai==0.2.14
langchain-anthropic==0.3.3
langchain-community==0.3.13
langchain-chroma==0.1.4
langgraph==0.2.59
chromadb==0.5.23
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.0
python-dotenv==1.0.1
httpx==0.27.2
python-multipart==0.0.20
tiktoken==0.8.0
tenacity==9.0.0
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-mock==3.14.0
langchain-ollama==0.2.0
ollama==0.4.4
